refactor: 아티스트 공연 일정 조회 객체 분리
