+++ backend/src/main/java/com/festago/artist/application/ArtistTotalSearchV1Service.java
@@ -3,7 +3,6 @@ package com.festago.artist.application;
 import com.festago.artist.dto.ArtistSearchStageCountV1Response;
 import com.festago.artist.dto.ArtistSearchTotalV1Response;
 import com.festago.artist.dto.ArtistSearchV1Response;
-import com.festago.artist.repository.ArtistV1SearchQueryDslRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
@@ -19,14 +18,14 @@ import org.springframework.transaction.annotation.Transactional;
 public class ArtistTotalSearchV1Service {
 
     private final ArtistSearchV1QueryService artistSearchV1QueryService;
-    private final ArtistV1SearchQueryDslRepository artistV1SearchQueryDslRepository;
+    private final ArtistSearchStageScheduleV1QueryService artistSearchStageScheduleV1QueryService;
 
     public List<ArtistSearchTotalV1Response> findAllByKeyword(String keyword, LocalDate today) {
         List<ArtistSearchV1Response> artists = artistSearchV1QueryService.findAllByKeyword(keyword);
         List<Long> artistIds = artists.stream()
             .map(ArtistSearchV1Response::id)
             .toList();
-        Map<Long, ArtistSearchStageCountV1Response> artistToStageCount = artistV1SearchQueryDslRepository.findArtistsStageScheduleAfterDateTime(
+        Map<Long, ArtistSearchStageCountV1Response> artistToStageCount = artistSearchStageScheduleV1QueryService.findArtistsStageScheduleAfterDateTime(
             artistIds, LocalDateTime.of(today, LocalTime.MIN));
         return artists.stream()
             .map(it -> ArtistSearchTotalV1Response.of(it, artistToStageCount.get(it.id())))
